requirements.txt
python-dotenv==1.0.1
requests==2.31.0
aiocron==1.8
beautifulsoup4==4.12.2
tmdbv3api==1.9.0
aligo~=6.2.2
python-telegram-bot[webhooks]==21.2
openai~=1.30.3
lxml==5.2.2
fake-useragent==1.5.1
httpx~=0.27.0